Have to refresh whole people list and update point bars
I was thinking update single person vote in people list box when
someone votes, this is working for people list box. However it's
not working for point bars, cuz point bars reads data from people
list box. So if people list box isn't updating, so is the point
bars. So, I have to use the previous way, to refresh people list
box, which is to update data in people list box, and then it in-
directly update point bars.